+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 2.9.5
 - phishing 0.18.5
 - api, api-contract, types 9.2.1
 - hw-ledger, keyring, util, util-crypto 10.1.4